requirements.txt
Eventlet==0.18.2
Flask==0.10.1
Flask-Admin==1.4.0
Flask-Login==0.2.11
Flask-Security==1.7.5
Flask-SocketIO==2.1
Flask-SQLAlchemy==2.1
Flask-Triangle==0.5.4
JsonPickle==0.9.2
